Containers no longer destroyed on shutdown
This was a bad idea. Since wharf is really just a management fronted to docker,
it was really rude to destroy everything in docker when the wharf shuts down.
There's no real reason to do be that disruptive to wharf users. The functionality
has been stubbed out, however, and may be added back with the future with more
flexibility.