eliminate docker container id -> name race condition
I was seeing errors in the log about docker being unable to retrieve the name
for a container immediately after getting a container list from docker. While
this is a race condition, there are also times when containers don't have names,
so at the expense of using the "pretty" names in the debug log, I've switched
the container sifter to only use the docker IDs, which has the side benefit
of bothering docker less.